multistomp: Switch to newer hotplug detection approach.
Using a spare GPIO to control hotplug detect is pointless since we can
simply pull D+ and D- to ground. Note that the USB cell on STM32F103
automatically grabs the GPIO pins when the transceiver is enabled so
there's no need to put the GPIO setting back to normal.

Signed-off-by: Daniel Thompson <daniel@redfelineninja.org.uk>